Improve quality of PRNG and match badge behavior.
Added scrambling of the 32 bit pseudorandom number to get 4 bit output
(instead of using low 4 bits). Changed PRNG random seeding to use
getrandom API. Added function level comments for the RNG module.

In consulting with Voja, it seems the manual was out of date. There is an
additional bit scrambling operation applied on the generated random number
before it is sent to the program. This is needed because congruential
pseudorandom number generators have very low periodicity in the low bits.
Simply taking the lowest 4 bits would result in very short cycles for the
generated pseudorandom numbers (period 16).

The revision 3 firmware had another bug that caused the implementation to
diverge from the documentation. The effective `a` constant for the PRNG was
0x838c4e6d instead of 0x41c64e6d (the high 16 bits were shifted one position
to the right, i.e. multiplied by two). This was corrected in revision 4. I
added a compile time constant -DFIRMWARE_V3 to allow using the old constant
if needed.