Add basic support for key presses.
Key up events cannot be detected with getch(), so for now only key presses are
supported. If a key is kept down, it will start emitting repeated events after
an initial delay and each will be treated like a new key press. AnyPress and
LastPress will be set the first time one of the recognized keys is pressed,
after which they won't be reset again. JustPress will be set each time a new
key is returned by getch(), and it will be reset when read by the program as
expected.