Alloc fixed sized program memory and loop program.
Changed struct program to use a fixed sized instruction array (4K instructions).
Loop PC back to start when reaching the end of the program memory and replace
error when PC points beyond the end of the program with an assertion (it should
not happen since PC cannot hold a value that exceeds addressable memory).

Although the badge program transferring protocol uses the length of the program
to avoid transferring useless bytes, at runtime the whole program memory is
always available and the program counter loops back to the beginning. Unused
program memory is initialized with zero, which is a no-op if ignoring the flags
("cp r0, 0").