Check return value of getrandom
Because getrandom has attribute warn_unused_result, compilation will
fail on some versions of GCC/glibc. According to the man page, this
should never fail because we're using non-blocking getrandom and
the seed size < 256 bytes, so add a simple loop that will call this until
we get the right size.

Tested on Ubuntu 22.04 in WSL, compiled fine and ran test app.

Fixes issue #5