Add "red mode" for LED display (#7)
Add "red mode" for LED display, start single step mode at PC 0.

This is activated with the -r switch. This makes the page
output show in red to emulate the red LEDs on the badge.

This also has a fix for single step mode to make it start
at PC 0, not PC 1.

Co-authored-by: Ben Combee <bcombee@roku.com>